+++ frontend/src/pages/landing/login/components/Email.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { FormControl, InputGroup } from "react-bootstrap";
+import { Mailbox } from 'react-bootstrap-icons';
+
+
+const Email = () => {
+  return (
+    <InputGroup>
+      <InputGroup.Text>
+        <Mailbox/>
+      </InputGroup.Text>
+      <FormControl
+        placeholder="Email"
+        type="text"
+      />
+    </InputGroup>
+  );
+};
+
+export default Email;